fixed rubocop offenses; deferred some offenses to .rubocop_todo.yml

+++ libraries/provider_delete_lines.rb
@@ -22,56 +22,56 @@
 require 'fileutils'
 require 'tempfile'
 
+# extends Chef class
 class Chef
+  # extends Provider class
   class Provider
+    # inherits from Chef::Provider
     class DeleteLines < Chef::Provider
-
       def load_current_resource
       end
 
       def action_edit
         regex = /#{new_resource.pattern}/
 
-        if ::File.exists?(new_resource.path) then
-          begin
-            f = ::File.open(new_resource.path, "r+")
+        return unless ::File.exist?(new_resource.path)
+        begin
+          f = ::File.open(new_resource.path, 'r+')
 
-            file_owner = f.lstat.uid
-            file_group = f.lstat.gid
-            file_mode = f.lstat.mode
+          file_owner = f.lstat.uid
+          file_group = f.lstat.gid
+          file_mode = f.lstat.mode
 
-            temp_file = Tempfile.new('foo')
+          temp_file = Tempfile.new('foo')
 
-            modified = false
+          modified = false
 
-            f.each_line do |line|
-              if line =~ regex then
-                modified = true
-              else
-                temp_file.puts line
-              end
+          f.each_line do |line|
+            if line =~ regex
+              modified = true
+            else
+              temp_file.puts line
             end
-
-            f.close
-
-            if modified then
-              temp_file.rewind
-              FileUtils.copy_file(temp_file.path,new_resource.path)
-              FileUtils.chown(file_owner,file_group,new_resource.path)
-              FileUtils.chmod(file_mode,new_resource.path)
-              new_resource.updated_by_last_action(true)
-            end
-
-          ensure
-            temp_file.close
-            temp_file.unlink
           end
-        end # ::File.exists
+
+          f.close
+
+          if modified
+            temp_file.rewind
+            FileUtils.copy_file(temp_file.path, new_resource.path)
+            FileUtils.chown(file_owner, file_group, new_resource.path)
+            FileUtils.chmod(file_mode, new_resource.path)
+            new_resource.updated_by_last_action(true)
+          end
+
+        ensure
+          temp_file.close
+          temp_file.unlink
+        end
       end # def action_edit
 
       def action_nothing
       end
-
     end
   end
 end
